added 101

+++ 0x0C-more_malloc_free/101-mul.c
@@ -8,8 +8,8 @@
  */
 int main(int argc, char *argv[])
 {
-	int i, j;
-	int a = 0, b = 0, product = 0;
+	unsigned long int i, j;
+	unsigned long int a = 0, b = 0, product = 0;
 	int aisneg = 0, bisneg = 0;
 	(void)argc;
